Working rough version of BufferHandler. Also fixed some warnings

+++ src/Camera.h
@@ -109,7 +109,7 @@ public:
     }
 
     // processes input received from a mouse scroll-wheel event. Only requires input on the vertical wheel-axis
-    void ProcessMouseScroll(float yoffset)
+    void ProcessMouseScroll(double yoffset)
     {
         Zoom -= (float)yoffset;
         if (Zoom < 1.0f)
